# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Channel Doctor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/title-optimizer:
    post:
      operationId: optimizeTitle
      tags: [ai]
      summary: Generate 10 optimized YouTube titles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TitleOptimizerBody"
      responses:
        "200":
          description: Optimized titles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleOptimizerResult"
  /ai/thumbnail-ab:
    post:
      operationId: thumbnailAbTest
      tags: [ai]
      summary: Compare two thumbnails using AI vision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThumbnailAbBody"
      responses:
        "200":
          description: Comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThumbnailAbResult"
  /ai/retention-mapper:
    post:
      operationId: retentionMapper
      tags: [ai]
      summary: Analyze a transcript for likely drop-off points
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetentionBody"
      responses:
        "200":
          description: Retention analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetentionResult"
  /ai/content-ideas:
    post:
      operationId: generateContentIdeas
      tags: [ai]
      summary: Generate content ideas for a channel/topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentIdeasBody"
      responses:
        "200":
          description: Ideas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentIdeasResult"
  /ai/why-failed:
    post:
      operationId: diagnoseVideo
      tags: [ai]
      summary: Diagnose why a specific video underperformed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhyFailedBody"
      responses:
        "200":
          description: Diagnosis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhyFailedResult"
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Conversational AI growth coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: Reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResult"
...
components:
  schemas:
...
    TitleOptimizerBody:
      type: object
      properties:
        currentTitle: { type: string, nullable: true }
        topic: { type: string, nullable: true }
        channelTitle: { type: string, nullable: true }
        audience: { type: string, nullable: true }
    OptimizedTitle:
      type: object
      properties:
        title: { type: string }
        ctrScore: { type: number }
        reasoning: { type: string }
        style: { type: string }
      required: [title, ctrScore, reasoning, style]
    TitleOptimizerResult:
      type: object
      properties:
        titles:
          type: array
          items: { $ref: "#/components/schemas/OptimizedTitle" }
        analysis: { type: string }
      required: [titles, analysis]
    ThumbnailAbBody:
      type: object
      properties:
        thumbnailA: { type: string }
        thumbnailB: { type: string }
        title: { type: string, nullable: true }
      required: [thumbnailA, thumbnailB]
    ThumbnailScoreBreakdown:
      type: object
      properties:
        thumbnail: { type: string, enum: [A, B] }
        ctrScore: { type: number }
        clarity: { type: number }
        emotion: { type: number }
        contrast: { type: number }
        textReadability: { type: number }
        notes: { type: string }
      required: [thumbnail, ctrScore, clarity, emotion, contrast, textReadability, notes]
    ThumbnailAbResult:
      type: object
      properties:
        winner: { type: string, enum: [A, B, tie] }
        confidence: { type: number }
        analysis: { type: string }
        breakdown:
          type: array
          items: { $ref: "#/components/schemas/ThumbnailScoreBreakdown" }
      required: [winner, confidence, analysis, breakdown]
    RetentionBody:
      type: object
      properties:
        videoId: { type: string, nullable: true }
        title: { type: string, nullable: true }
        durationSeconds: { type: number, nullable: true }
    RetentionCurvePoint:
      type: object
      properties:
        secondsIn: { type: integer }
        estimatedRetentionPct: { type: number }
      required: [secondsIn, estimatedRetentionPct]
    RetentionDropoff:
      type: object
      properties:
        startSec: { type: integer }
        endSec: { type: integer }
        severity: { type: string, enum: [minor, major, critical] }
        cause: { type: string }
        fix: { type: string }
      required: [startSec, endSec, severity, cause, fix]
    RetentionResult:
      type: object
      properties:
        summary: { type: string }
        curve:
          type: array
          items: { $ref: "#/components/schemas/RetentionCurvePoint" }
        dropoffs:
          type: array
          items: { $ref: "#/components/schemas/RetentionDropoff" }
        hookAdvice: { type: string }
      required: [summary, curve, dropoffs, hookAdvice]
    ContentIdeasBody:
      type: object
      properties:
        topic: { type: string, minLength: 1 }
        channelTitle: { type: string, nullable: true }
        audience: { type: string, nullable: true }
      required: [topic]
    ContentIdea:
      type: object
      properties:
        title: { type: string }
        hook: { type: string }
        format: { type: string }
        why: { type: string }
      required: [title, hook, format, why]
    ContentIdeasResult:
      type: object
      properties:
        ideas:
          type: array
          items: { $ref: "#/components/schemas/ContentIdea" }
      required: [ideas]
    WhyFailedBody:
      type: object
      properties:
        videoId: { type: string, minLength: 1 }
        channelId: { type: string, nullable: true }
      required: [videoId]
    WhyFailedReason:
      type: object
      properties:
        category: { type: string }
        severity: { type: string, enum: [minor, major, critical] }
        explanation: { type: string }
      required: [category, severity, explanation]
    WhyFailedFix:
      type: object
      properties:
        action: { type: string }
        impact: { type: string }
      required: [action, impact]
    WhyFailedResult:
      type: object
      properties:
        verdict: { type: string }
        gapVsAverage: { type: number }
        reasons:
          type: array
          items: { $ref: "#/components/schemas/WhyFailedReason" }
        fixes:
          type: array
          items: { $ref: "#/components/schemas/WhyFailedFix" }
        titleAlternatives:
          type: array
          items: { type: string }
      required: [verdict, gapVsAverage, reasons, fixes, titleAlternatives]
...
    AiChatMessage:
      type: object
      properties:
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
      required: [role, content]
    AiChatRequest:
      type: object
      properties:
        message: { type: string }
        channelId: { type: string }
        history:
          type: array
          items: { $ref: "#/components/schemas/AiChatMessage" }
      required: [message]
    AiChatResult:
      type: object
      properties:
        reply: { type: string }
      required: [reply]
...